Docker Compose: Eclipse Mosquitto
Eclipse Mosquitto is an open source (EPL/EDL licensed) message broker that implements
the MQTT protocol versions 3.1 and 3.1.1. Mosquitto is lightweight and is suitable for
use on all devices from low power single board computers to full servers.

The MQTT protocol provides a lightweight method of carrying out messaging using a
publish/subscribe model. This makes it suitable for Internet of Things messaging
such as with low power sensors or mobile devices such as phones, embedded computers
or microcontrollers.

https://mosquitto.org/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>