Home Assistant: FFmpeg Camera Component
The ffmpeg platform allows you to use any video feed as a camera in
Home Assistant via FFmpeg. This video source must support multiple
simultaneous reads, because for every concurrent Home Assistant user,
a connection will be made to the source every 10 seconds. Normally
this should not be a problem.

https://www.home-assistant.io/components/camera.ffmpeg/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>