Home Assistant: Dark Sky Sensor Component
The darksky platform uses the Dark Sky web service as a source for
meteorological data for your location. The location is based on the
longitude and latitude coordinates configured in your
configuration.yaml file.

The coordinates are auto-detected but to take advantage of the
hyper-local weather reported by Dark Sky, you can refine them down
to your exact home address.

https://www.home-assistant.io/components/sensor.darksky/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>