Docker Compose: Node-RED
Node-RED is a programming tool for wiring together hardware devices,
APIs and online services in new and interesting ways.

It provides a browser-based editor that makes it easy to wire together
flows using the wide range of nodes in the palette that can be deployed
to its runtime in a single-click.

https://nodered.org/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>