Home Assistant: Configuration Storing Secrets
The configuration.yaml file is a plain-text file, thus it is readable by
anyone who has access to the file. The file contains passwords and API tokens
which need to be redacted if you want to share your configuration. By using
!secret you can remove any private information from your configuration files.
This separation can also help you to keep easier track of your passwords and
API keys, as they are all stored at one place and no longer spread across the
configuration.yaml file or even multiple yaml files if you split up your
configuration.

Make this the default way to manage URLs to point to the rigt IP address of a
our services, specially in remote operations.

https://www.home-assistant.io/docs/configuration/secrets/

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>