OpenVINO: Object Size Detector Python
This Object Size Detector application is one of a series of computer vision
(CV) reference implementations using the Intel® Distribution of OpenVINO™
toolkit. This application demonstrates how to use CV to detect and measure the
approximate length, width and size of assembly line parts. It is designed to
work with an assembly line camera mounted above the assembly line belt. The
application monitors mechanical parts as they are moving down the assembly
line and raises an alert if it detects a part on the belt outside a specified
size range.

https://github.com/intel-iot-devkit/object-size-detector-python

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>