OpenVINO: Store Traffic Monitor Python
This Store Traffic Monitor is one of a series of IoT reference implementations
aimed at instructing users on how to develop a working solution for a
particular problem. It demonstrates how to create a smart video IoT solution
using Intel® hardware and software tools. This reference implementation
monitors people activity inside and outside a facility as well as counts
product inventory.

https://github.com/intel-iot-devkit/store-traffic-monitor-python

Signed-off-by: Abraham Arce <xe1gyq@gmail.com>
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>